+++ app/src/main/java/com/murzify/horseracing/core/domain/model/RaceResult.kt
@@ -6,5 +6,5 @@ data class RaceResult(
     val id: Int,
     val winner: Int,
     val time: LocalDateTime,
-    val durationMillis: Long
+    val durationMillis: Int
 )
